Link to the full contribution guide
GitHub pushes projects to have a contribution guide in a file called
CONTRIBUTING.md under their respective code repositories. This is
reflected in the project's Insights -> Community section.

Hence, this guide can't be fully moved to the website.

Instead, to avoid duplicating the content, it's better to link to the
contribution guide from the website.

https://github.com/containers/containertoolbx.org/pull/12