Put the rest of the earlier introductory text in manual.md
... and slightly shorten the existing text to accommodate it.

The second-level manual.md isn't as prime a real estate as the top-level
index.md, and Toolbx is too closely associated with Fedora to not
highlight how it fits with Fedora Silverblue and Workstation.

Secondly, the earlier introductory text was carefully massaged in
consultation with various stakeholders over the years to strike the
right balance between making it appealing and understable to the casual
reader, and technical accuracy and details for those who are intimately
familiar with Toolbx.  Removing it without careful thought can re-ignite
needless debates about topics like security that are best avoided right
now [1].

[1] https://github.com/containers/toolbox/issues/183
    https://github.com/containers/toolbox/issues/1020