Reduce the use of the word 'support' and clarify what it means
Readers sometimes misinterpret support for a set of operating systems to
mean that Toolbx is not expected to work elsewhere and that they
shouldn't even try it.  This is too harsh and negative.

In reality, Toolbx is regularly tested with a certain combination of
host OSes and OCI images, and those are expected to work well.  Even
though other combinations may or may not work, the goal is not to
discourage users from trying them out.

As more and more hosts and images are added to the test matrix, it
increases the possibility that those outside the matrix will also work.
eg., Toolbx is tested on Fedora and Red Hat Enterprise Linux, and there
is a good chance that it will work on the Red Hat In-Vehicle Operating
System (or RHIVOS).

Regularly testing Toolbx across a combination of hosts and images is a
matter of CI and human maintenance resources, and popular demand.  It
can never cover every possible combination, and the set will ebb and
flow as those factors change over time.

The text about Windows Subsystem for Linux was left out because its not
clear if Toolbx works there, and it might add to the confusion.