Mention access to the host file system at /run/host
The ability to access the host file system from inside the container at
/run/host is an important feature that lets users add their own custom
bind mounts and symbolic links.  It emphasizes that Toolbx makes no
promise about security beyond what's already available in the usual
command line environment on the host.

Therefore, it's good to mention it prominently outside the manual page
for the 'create' command [1].

See:
https://github.com/containers/toolbox/commit/d03a5fee80f2f72d
https://github.com/containers/toolbox/pull/827

[1] https://github.com/containers/toolbox/blob/main/doc/toolbox-create.1.md

https://github.com/containers/toolbox/issues/1649