Clarify how Toolbx environments are separate from the host
It's already been mentioned a few times that Toolbx environments have
seamless access to the user’s home directory.  It's worth reminding the
reader that their /usr file systems are separate from the host.  Else,
it might not be clear how or why the development environments are
detached from the host operating system.

https://github.com/containers/toolbox/issues/1649