Add text on paths under image requirements
The wording was changed from the original text before commit
d1f48731ec76a3ae to not mention bind mounts.  The concern about bind
mounts had more to do with unexpected changes in or wrong use of Podman,
and less to do with how the images were created.

Instead, the text says what kind of files these paths must be, which is
more useful for someone creating Toolbx images.

https://github.com/containers/toolbox/issues/1606