Tweak the text in the PKCS #11 sub-section for clarity and consistency
It wasn't clear if the /etc/pkcs11/modules directory should be present
if the images are configured to use it and that it should be possible to
create a file inside it.

There's no need to put the suggestion for filing an issue in a separate
paragraph.  Combining the paragraph shortens the text and structures it
as elsewhere.

Fallout from 8f8bf407e416bb6e36751161acf20180e3c6d1e8

https://github.com/containers/toolbox/issues/1606