Unify 'Commands' and 'Paths', and rename the combined sub-section
Many individual commands and paths have their own sub-sections now to
make it possible to add more details specific to them or to keep the
requirements around a specific feature in the same place.  Therefore,
the 'commands' and 'paths' sub-sections are now shorter than they used
to be.

The combined sub-section was renamed so that it gets alphabetically
ordered as the last one.

This reverts commit 7c4666a43ee812aba9aa001d0b7dc3be3276df86.

https://github.com/containers/toolbox/issues/1606